Avance

+++ Proyecto.DAL/ModelosRepositorios/UsuarioRepository.cs
@@ -2,10 +2,7 @@
 using Proyecto.DAL.DataContext;
 using Proyecto.DAL.Repositorio;
 using Proyecto.MODELS;
-using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Proyecto.DAL.ModelosRepositorios
@@ -13,31 +10,32 @@ namespace Proyecto.DAL.ModelosRepositorios
     public class UsuarioRepository : IGenericRepository<Usuario>
     {
         private readonly PuntoVentaContext _dbContext;
-        public UsuarioRepository (PuntoVentaContext context)
+
+        public UsuarioRepository(PuntoVentaContext context)
         {
             _dbContext = context;
+        }
+
+        public async Task<bool> Crear(Usuario modelo)
+        {
+            await _dbContext.Usuarios.AddAsync(modelo);
+            return await _dbContext.SaveChangesAsync() > 0;
         }
 
         public async Task<bool> Actualizar(Usuario modelo)
         {
             _dbContext.Usuarios.Update(modelo);
-            await _dbContext.SaveChangesAsync();
-            return true;
-        }
-
-        public async Task<bool> Crear(Usuario modelo)
-        {
-            _dbContext.Usuarios.Add(modelo);
-            await _dbContext.SaveChangesAsync();
-            return true;
+            return await _dbContext.SaveChangesAsync() > 0;
         }
 
         public async Task<bool> Eliminar(int id)
         {
-            Usuario modelo = _dbContext.Usuarios.First(c => c.IdUsuario == id);
+            var modelo = await _dbContext.Usuarios.FirstOrDefaultAsync(c => c.IdUsuario == id);
+            if (modelo == null)
+                return false;
+
             _dbContext.Usuarios.Remove(modelo);
-            await _dbContext.SaveChangesAsync();
-            return true;
+            return await _dbContext.SaveChangesAsync() > 0;
         }
 
         public async Task<Usuario> obtener(int id)
@@ -45,10 +43,9 @@ namespace Proyecto.DAL.ModelosRepositorios
             return await _dbContext.Usuarios.FindAsync(id);
         }
 
-        public async Task<IQueryable<Usuario>> obtenerTodos()
+        public Task<IQueryable<Usuario>> obtenerTodos()
         {
-            IQueryable<Usuario> queryUsuario = _dbContext.Usuarios;
-            return queryUsuario;
+            return Task.FromResult(_dbContext.Usuarios.AsQueryable());
         }
     }
 }
